requirements.txt
eth-account
x402
python-dotenv
opencv-python
pyaudio
numpy
openai
elevenlabs